requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
pydantic==2.10.4
pydantic-settings==2.7.1

# Image processing (stubbed renderer)
Pillow==11.1.0

# AWS SDK
boto3==1.36.7

# Utilities
python-jose[cryptography]==3.3.0
httpx==0.28.1
aiofiles==24.1.0

# AWS CDK
aws-cdk-lib==2.177.0
constructs==10.4.2
